utils: move out of internal
